# swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"ISample": {
			"properties": {
				"name": {
					"type": "string"
				},
				"url": {
					"type": "string"
				},
				"isActive": {
					"type": "boolean"
				}
			},
			"required": [
				"name",
				"url",
				"isActive"
			],
			"type": "object",
			"additionalProperties": false
		}
	},
	"info": {
		"title": "ts-node-boilerplate",
		"version": "1.0.0",
		"description": "app platform backend",
		"license": {
			"name": "ISC"
		}
	},
	"paths": {
		"/samples": {
			"get": {
				"operationId": "List",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"items": {
								"$ref": "#/definitions/ISample"
							},
							"type": "array"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "filter",
						"required": false,
						"type": "string"
					}
				]
			},
			"post": {
				"operationId": "Create",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/ISample"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "plan",
						"required": true,
						"schema": {
							"$ref": "#/definitions/ISample"
						}
					}
				]
			}
		},
		"/samples/{id}": {
			"get": {
				"operationId": "Find",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/ISample"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"format": "double",
						"type": "number"
					}
				]
			},
			"put": {
				"operationId": "Update",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"additionalProperties": true,
							"type": "object"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"format": "double",
						"type": "number"
					},
					{
						"in": "body",
						"name": "plan",
						"required": true,
						"schema": {
							"$ref": "#/definitions/ISample"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"additionalProperties": true,
							"type": "object"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"format": "double",
						"type": "number"
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {}
}